+++ src/components/SInput/index.ts
@@ -1,1 +1,2 @@
 export { default as SInput } from "./SInput.vue";
+export * from "./SInput.vue";