+++ src/components/SInputGroup/SInputBlock.stories.ts
@@ -1,0 +1,279 @@
+import SInputBlock from './SInputBlock.vue';
+import { ref } from 'vue';
+import { SButton } from '../SButton';
+import { SDropdown, SDropdownItem } from '../SDropdown';
+import type { SourceProps } from '@storybook/blocks';
+import { buildDesign, buildSourceBinding } from '../../helpers';
+import {
+  ArrowLeftOnRectangleIcon,
+  InformationCircleIcon,
+  CurrencyDollarIcon,
+  MapPinIcon,
+  MagnifyingGlassIcon,
+} from '@heroicons/vue/24/outline';
+import { EnvelopeIcon, KeyIcon, ChatBubbleLeftEllipsisIcon } from '@heroicons/vue/24/solid';
+
+export default {
+  component: SInputBlock,
+  title: 'new/InputBlock',
+  parameters: {
+    docs: {
+      description: {
+        component: 'DOC',
+      },
+    },
+  },
+  argTypes: {
+    // Events
+    'update:modelValue': {
+      control: { type: null },
+      table: { type: { summary: null }, category: 'Events' },
+      description: 'DOC',
+    },
+
+    // Props
+    error: {
+      control: 'text',
+      description: 'DOC',
+      table: { type: { summary: 'string' }, category: 'Props' },
+    },
+    helpText: {
+      control: 'text',
+      description: 'DOC',
+      table: { type: { summary: 'string' }, category: 'Props' },
+    },
+    label: {
+      control: 'text',
+      description: 'DOC',
+      table: { type: { summary: 'string' }, category: 'Props' },
+    },
+
+    // Props - Input
+    disabled: {
+      description: 'DOC',
+      table: { type: { summary: 'boolean' }, category: 'Props', subcategory: 'Input' },
+    },
+    endIcon: {
+      control: 'select',
+      options: [
+        'EnvelopeIcon',
+        'MagnifyingGlassIcon',
+        'MapPinIcon',
+        'InformationCircleIcon',
+        'CurrencyDollarIcon',
+        'ChatBubbleLeftEllipsisIcon',
+      ],
+      description: `A Vue functional component to be used as the input's icon.`,
+      table: { type: { summary: 'FunctionalComponent' }, category: 'Props', subcategory: 'Input' },
+    },
+    icon: {
+      control: 'select',
+      options: [
+        'EnvelopeIcon',
+        'MagnifyingGlassIcon',
+        'MapPinIcon',
+        'InformationCircleIcon',
+        'CurrencyDollarIcon',
+        'ChatBubbleLeftEllipsisIcon',
+      ],
+      description: `A Vue functional component to be used as the input's icon.`,
+      table: { type: { summary: 'FunctionalComponent' }, category: 'Props', subcategory: 'Input' },
+    },
+    id: {
+      control: 'text',
+      description: 'DOC',
+      table: { type: { summary: 'string' }, category: 'Props', subcategory: 'Input' },
+    },
+    modelValue: {
+      control: { type: null },
+      description: 'DOC',
+      table: { type: { summary: 'Ref<string>' }, category: 'Props', subcategory: 'Input' },
+    },
+    name: {
+      control: 'text',
+      description: 'DOC',
+      table: { type: { summary: 'string' }, category: 'Props', subcategory: 'Input' },
+    },
+    rounded: {
+      control: 'inline-radio',
+      options: ['left', 'right', 'both', 'none'],
+      description: `Specifies which corners should be rounded.`,
+      table: { type: { summary: 'left | right | both | none | full' }, category: 'Props', subcategory: 'Input' },
+    },
+    placeholder: {
+      control: 'text',
+      description: 'DOC',
+      table: { type: { summary: 'string' }, category: 'Props', subcategory: 'Input' },
+    },
+    prefix: {
+      control: 'text',
+      description: 'DOC',
+      table: { type: { summary: 'string' }, category: 'Props', subcategory: 'Input' },
+    },
+    suffix: {
+      control: 'text',
+      description: 'DOC',
+      table: { type: { summary: 'string' }, category: 'Props', subcategory: 'Input' },
+    },
+    type: {
+      control: 'text',
+      description: 'DOC',
+      table: { type: { summary: 'string' }, category: 'Props', subcategory: 'Input' },
+    },
+  },
+};
+
+const design = buildDesign('');
+
+const sourceBinding = buildSourceBinding({
+  prop: {
+    rounded: 'both',
+    id: undefined,
+    name: undefined,
+    placeholder: undefined,
+    prefix: undefined,
+    suffix: undefined,
+    type: 'text',
+  },
+  check: ['disabled', 'error'],
+  custom: { icon: true },
+});
+
+export const Default = {
+  render: (args: any) => ({
+    components: {
+      SInputBlock,
+      EnvelopeIcon,
+      MagnifyingGlassIcon,
+      MapPinIcon,
+      InformationCircleIcon,
+      CurrencyDollarIcon,
+      ChatBubbleLeftEllipsisIcon,
+    },
+    setup() {
+      const value = ref('');
+      const getIcon = (
+        icon:
+          | 'EnvelopeIcon'
+          | 'MagnifyingGlassIcon'
+          | 'MapPinIcon'
+          | 'InformationCircleIcon'
+          | 'CurrencyDollarIcon'
+          | 'ChatBubbleLeftEllipsisIcon'
+      ) => {
+        if (icon === 'EnvelopeIcon') return EnvelopeIcon;
+        if (icon === 'MagnifyingGlassIcon') return MagnifyingGlassIcon;
+        if (icon === 'MapPinIcon') return MapPinIcon;
+        if (icon === 'InformationCircleIcon') return InformationCircleIcon;
+        if (icon === 'CurrencyDollarIcon') return CurrencyDollarIcon;
+        if (icon === 'ChatBubbleLeftEllipsisIcon') return ChatBubbleLeftEllipsisIcon;
+      };
+
+      return {
+        args,
+        getIcon,
+        value,
+        EnvelopeIcon,
+        MagnifyingGlassIcon,
+        MapPinIcon,
+        InformationCircleIcon,
+        CurrencyDollarIcon,
+        ChatBubbleLeftEllipsisIcon,
+      };
+    },
+    template: '<SInputBlock v-bind="args" :icon="getIcon(args.icon)" :endIcon="getIcon(args.endIcon)" v-model="value" />',
+  }),
+  parameters: {
+    design,
+    docs: {
+      canvas: { layout: 'centered' },
+      source: {
+        transform: ((_, storyContext) => `
+        <SInputBlock ${sourceBinding(storyContext.args)} />
+        `) as SourceProps['transform'],
+        type: 'dynamic',
+        language: 'html',
+      },
+    },
+  },
+  args: {
+    disabled: false,
+    error: '',
+    helpText: 'Help text',
+    id: 'test-id',
+    label: 'Label',
+    name: 'test-name',
+    placeholder: 'Placeholder',
+    prefix: '',
+    suffix: '',
+    rounded: 'both',
+    type: 'text',
+  },
+};
+
+const createVariation = (
+  template: string,
+  options?: {
+    focusVisible?: boolean;
+    containerClass?: string;
+  }
+) => ({
+  decorators: [
+    () => ({ template: `<div style="${options?.containerClass ?? 'gap: 20px; display: flex;'}"><story/></div>` }),
+  ],
+  render: () => ({
+    components: {
+      SInputBlock,
+      SButton,
+      SDropdown,
+      SDropdownItem,
+      ArrowLeftOnRectangleIcon,
+      EnvelopeIcon,
+      KeyIcon,
+      InformationCircleIcon,
+      ChatBubbleLeftEllipsisIcon,
+      CurrencyDollarIcon,
+      MapPinIcon,
+    },
+    setup() {
+      const email = ref('');
+
+      return {
+        email,
+        SButton,
+        ArrowLeftOnRectangleIcon,
+        EnvelopeIcon,
+        KeyIcon,
+        InformationCircleIcon,
+        ChatBubbleLeftEllipsisIcon,
+        CurrencyDollarIcon,
+        MapPinIcon,
+      };
+    },
+    template,
+  }),
+  parameters: {
+    design,
+    pseudo: { focusVisible: options?.focusVisible },
+    controls: { disable: true },
+    actions: { disable: true },
+    docs: {
+      source: {
+        code: template,
+        language: 'html',
+      },
+    },
+  },
+});
+
+export const WithLabel = createVariation(`<SInputBlock label="Name" placeholder="Enter your name" />`, {
+  containerClass: 'width: 200px',
+});
+
+export const WithHelpText = createVariation(`<SInputBlock label="Token" helpText="does not include quotes" placeholder="XX-XXXX-XXXX" />`, {
+  containerClass: 'width: 200px',
+});
+
+export const WithErrorText = createVariation(`<SInputBlock label="Amount" error="the amount is mandatory" prefix="$" placeholder="0.00" />`, {
+  containerClass: 'width: 200px',
+});