+++ src/components/SInputGroup/index.ts
@@ -1,0 +1,1 @@
+export { default as SInputBlock } from "./SInputBlock.vue";